+++ appBackend/appBackend/build.gradle.kts
@@ -2,17 +2,12 @@ import com.gchristov.thecodinglove.gradleplugins.Deps
 
 plugins {
     id("backend-executable-plugin")
-    // TODO: This shouldn't be needed here but is used for examples
-    id("kmp-data-plugin")
 }
 
 kotlin {
     sourceSets {
         val jsMain by getting {
             dependencies {
-                // TODO: Use feature modules here when ready. These are only used for examples
-                implementation(projects.kmpCommonKotlin)
-                implementation(projects.search)
                 implementation(projects.slack)
                 // Ideally these would be linked from corresponding submodules but that is currently
                 // not supported out of the box or through the npm-publish plugin and causes "module
