+++ modules.gradle.kts
@@ -7,6 +7,7 @@ val projects = listOf(
     "kmp/kmp-common-network",
     "kmp/kmp-common-test",
     "appBackend/appBackend",
+    "appBackend/common-firebase",
     "appBackend/htmlparse",
     "appBackend/htmlparse-data",
     "appBackend/htmlparse-testfixtures",
